+++ mshellctl/src/subcommands/mod.rs
@@ -1,5 +1,6 @@
 pub mod audio;
 pub mod bar;
+pub mod bluetooth;
 pub mod brightness;
 pub mod lock;
 pub mod media;